sort table files
