+++ packages/client/src/pages/MainPage/containers/StatementEditorBox/StatementEditorActantTable/StatementEditorActantTableStyles.tsx
@@ -28,7 +28,7 @@ export const StyledTr = styled.tr<StyledTr>`
     padding-left: ${({ theme }) => theme.space[1]};
     padding-right: ${({ theme }) => theme.space[2]};
   }
-  td:not(:last-child) {
+  > td {
     width: 1%;
   }
 `;
